+++ src/time_parsers/time_parsers.ts
@@ -3,6 +3,7 @@ import { toFixed, toFixedFloat } from '../number_parsers/index';
 
 // Installed Package
 import VDate from '@vieolo/date';
+import { time24Hours, time24HoursWithSeconds } from '@vieolo/regex-library';
 
 export function parseHourAndMinute(hour: number, minute: number) : string;
 export function parseHourAndMinute(hour: number, minute: number, returnType?: 'string') : string;
@@ -54,4 +55,52 @@ export function parseDecimalTime(decimal: number, includeSecond?: boolean): stri
 	let outputSecond = `${finalSeconds < 10 ? "0" + finalSeconds : finalSeconds}`;
 
 	return `${outputHour}:${outputMinute}${includeSecond ? ":" + outputSecond : ""}`;
+}
+
+
+/** parses 126 to 2.10 (string) */
+export function parseMinuteCount(count: number, returnType: 'decimal-string') : string;
+/** parses 126 to 2.10 (number) */
+export function parseMinuteCount(count: number, returnType: 'decimal-number') : number;
+/** parses 126 to 2h 6m */
+export function parseMinuteCount(count: number, returnType: 'formatted-hour-minute') : string;
+/** parses 126 to 02:06 */
+export function parseMinuteCount(count: number, returnType: 'formatted-time') : string;
+/** parses 126 to a VDate instance with the current day, month, and year, at 02:06 AM */
+export function parseMinuteCount(count: number, returnType: 'vdate') : VDate;
+export function parseMinuteCount(count: number, returnType: 'decimal-string' | 'decimal-number' | 'formatted-hour-minute' | 'formatted-time' | 'vdate') : any {
+	if (returnType == 'decimal-number') return toFixedFloat(count / 60, 2);
+	else if (returnType == 'decimal-string') return toFixed(count / 60, 2);
+	else if (returnType == 'formatted-hour-minute') return VDate.getTimeFromMinuteCount(count);
+
+	let hourCount = Math.floor(count / 60);
+	let minuteCount = count % 60;
+
+	if (returnType == 'formatted-time') return parseHourAndMinute(hourCount, minuteCount, 'string');
+	else return parseHourAndMinute(hourCount, minuteCount, 'vdate');
+}
+
+
+/** parses 02:06 to 2.10 (number) */
+export function parseFormattedTime(time: string, returnType: 'decimal-number') : number;
+/** parses 02:06 to 2.10 (string) */
+export function parseFormattedTime(time: string, returnType: 'decimal-string') : string;
+export function parseFormattedTime(time: string, returnType: 'decimal-number' | 'decimal-string') : any {
+	
+	if (!time24Hours.test(time) && !time24HoursWithSeconds.test(time)) throw new Error("The given time is not valid");
+	
+	let splitted = time.split(':');
+
+	let minuteCount = (
+		parseInt((splitted[0])) * 60) +
+		parseInt(splitted[1]) +
+		((splitted.length > 2 && parseInt(splitted[2]) > 29) ? 1 : 0
+	) // Adding one minute if the duration has any seconds
+	
+	let parsed: string | number;
+	
+	if (returnType == 'decimal-number') parsed = parseMinuteCount(minuteCount, 'decimal-number');
+	else parsed = parseMinuteCount(minuteCount, 'decimal-string');
+	
+    return parsed;
 }