+++ tests/time_parsers.test.ts
@@ -1,5 +1,10 @@
 // Parsers
-import { parseHourAndMinute, parseDecimalTime } from '../src/time_parsers/time_parsers'
+import { 
+    parseHourAndMinute, 
+    parseDecimalTime,
+    parseMinuteCount,
+    parseFormattedTime
+} from '../src/time_parsers/time_parsers'
 
 
 describe("Time Parsers", () => {
@@ -583,6 +588,154 @@ describe("Time Parsers", () => {
         expect(parseDecimalTime(67.35, true)).toEqual("67:21:00")
         expect(parseDecimalTime(2.05, true)).toEqual("02:03:00")
         expect(parseDecimalTime(8.05, true)).toEqual("08:03:00")
+
+    })
+
+
+    it('parses minute count correctly', () => {
+
+        //#region to decimal-string
+
+        expect(parseMinuteCount(126, 'decimal-string')).toBe("2.10");
+        expect(parseMinuteCount(0, 'decimal-string')).toBe("0.00");
+        expect(parseMinuteCount(1, 'decimal-string')).toBe("0.02");
+        expect(parseMinuteCount(60, 'decimal-string')).toBe("1.00");
+        expect(parseMinuteCount(76, 'decimal-string')).toBe("1.27");
+
+        //#endregion
+
+
+        //#region to decimal-number
+
+        expect(parseMinuteCount(126, 'decimal-number')).toBe(2.10);
+        expect(parseMinuteCount(0, 'decimal-number')).toBe(0.00);
+        expect(parseMinuteCount(1, 'decimal-number')).toBe(0.02);
+        expect(parseMinuteCount(60, 'decimal-number')).toBe(1.00);
+        expect(parseMinuteCount(76, 'decimal-number')).toBe(1.27);
+
+        //#endregion
+
+
+
+        //#region to formatted-hour-minute
+
+        expect(parseMinuteCount(126, 'formatted-hour-minute')).toBe("2h 6m");
+        expect(parseMinuteCount(0, 'formatted-hour-minute')).toBe("0m");
+        expect(parseMinuteCount(1, 'formatted-hour-minute')).toBe("1m");
+        expect(parseMinuteCount(60, 'formatted-hour-minute')).toBe("1h 0m");
+        expect(parseMinuteCount(76, 'formatted-hour-minute')).toBe("1h 16m");
+
+        //#endregion
+
+
+        //#region to formatted-time
+
+        expect(parseMinuteCount(126, 'formatted-time')).toBe("02:06");
+        expect(parseMinuteCount(0, 'formatted-time')).toBe("00:00");
+        expect(parseMinuteCount(1, 'formatted-time')).toBe("00:01");
+        expect(parseMinuteCount(60, 'formatted-time')).toBe("01:00");
+        expect(parseMinuteCount(76, 'formatted-time')).toBe("01:16");
+
+        //#endregion
+
+
+        //#region to VDate
+
+        expect(parseMinuteCount(126, 'vdate').getHours()).toBe(2);
+        expect(parseMinuteCount(126, 'vdate').getMinutes()).toBe(6);
+        expect(parseMinuteCount(0, 'vdate').getHours()).toBe(0);
+        expect(parseMinuteCount(0, 'vdate').getMinutes()).toBe(0);
+        expect(parseMinuteCount(1, 'vdate').getHours()).toBe(0);
+        expect(parseMinuteCount(1, 'vdate').getMinutes()).toBe(1);
+        expect(parseMinuteCount(60, 'vdate').getHours()).toBe(1);
+        expect(parseMinuteCount(60, 'vdate').getMinutes()).toBe(0);
+        expect(parseMinuteCount(76, 'vdate').getHours()).toBe(1);
+        expect(parseMinuteCount(76, 'vdate').getMinutes()).toBe(16);
+
+        //#endregion
+
+    })
+
+
+
+    it("parses formatted time correctly", () => {
+
+        //#region to decimal-number
+
+        expect(parseFormattedTime("11:57", 'decimal-number')).toEqual(11.95);
+        expect(parseFormattedTime("12:07", 'decimal-number')).toEqual(12.12);
+        expect(parseFormattedTime("10:51", 'decimal-number')).toEqual(10.85);
+        expect(parseFormattedTime("14:12", 'decimal-number')).toEqual(14.2);
+        expect(parseFormattedTime("10:38", 'decimal-number')).toEqual(10.63);
+        expect(parseFormattedTime("08:26", 'decimal-number')).toEqual(8.43);
+        expect(parseFormattedTime("00:00", 'decimal-number')).toEqual(0);
+        expect(parseFormattedTime("11:53", 'decimal-number')).toEqual(11.88);
+        expect(parseFormattedTime("12:27", 'decimal-number')).toEqual(12.45);
+        expect(parseFormattedTime("09:54", 'decimal-number')).toEqual(9.9);
+        expect(parseFormattedTime("12:38", 'decimal-number')).toEqual(12.63);
+        expect(parseFormattedTime("11:30", 'decimal-number')).toEqual(11.50);
+        expect(parseFormattedTime("09:05", 'decimal-number')).toEqual(9.08);
+        //expect(parseFormattedTime("67:21", 'decimal-number')).toThrowError("The given time is not valid")
+        expect(parseFormattedTime("02:03", 'decimal-number')).toEqual(2.05)
+        expect(parseFormattedTime("08:03", 'decimal-number')).toEqual(8.05)
+
+        expect(parseFormattedTime("11:57:00", 'decimal-number')).toEqual(11.95);
+        expect(parseFormattedTime("12:07:12", 'decimal-number')).toEqual(12.12);
+        expect(parseFormattedTime("10:51:00", 'decimal-number')).toEqual(10.85);
+        expect(parseFormattedTime("14:12:00", 'decimal-number')).toEqual(14.2);
+        expect(parseFormattedTime("10:37:48", 'decimal-number')).toEqual(10.63);
+        expect(parseFormattedTime("08:25:48", 'decimal-number')).toEqual(8.43);
+        expect(parseFormattedTime("00:00:00", 'decimal-number')).toEqual(0);
+        expect(parseFormattedTime("11:52:48", 'decimal-number')).toEqual(11.88);
+        expect(parseFormattedTime("12:27:00", 'decimal-number')).toEqual(12.45);
+        expect(parseFormattedTime("09:54:00", 'decimal-number')).toEqual(9.9);
+        expect(parseFormattedTime("12:37:48", 'decimal-number')).toEqual(12.63);
+        expect(parseFormattedTime("11:30:00", 'decimal-number')).toEqual(11.50);
+        expect(parseFormattedTime("09:04:48", 'decimal-number')).toEqual(9.08);
+        //expect(parseFormattedTime("67:21:00", 'decimal-number')).toThrow(new Error("The given time is not valid"))
+        expect(parseFormattedTime("02:03:00", 'decimal-number')).toEqual(2.05)
+        expect(parseFormattedTime("08:03:00", 'decimal-number')).toEqual(8.05)
+
+        //#endregion
+
+
+        //#region to decimal-string
+
+        expect(parseFormattedTime("11:57", 'decimal-string')).toEqual("11.95");
+        expect(parseFormattedTime("12:07", 'decimal-string')).toEqual("12.12");
+        expect(parseFormattedTime("10:51", 'decimal-string')).toEqual("10.85");
+        expect(parseFormattedTime("14:12", 'decimal-string')).toEqual("14.20");
+        expect(parseFormattedTime("10:38", 'decimal-string')).toEqual("10.63");
+        expect(parseFormattedTime("08:26", 'decimal-string')).toEqual("8.43");
+        expect(parseFormattedTime("00:00", 'decimal-string')).toEqual("0.00");
+        expect(parseFormattedTime("11:53", 'decimal-string')).toEqual("11.88");
+        expect(parseFormattedTime("12:27", 'decimal-string')).toEqual("12.45");
+        expect(parseFormattedTime("09:54", 'decimal-string')).toEqual("9.90");
+        expect(parseFormattedTime("12:38", 'decimal-string')).toEqual("12.63");
+        expect(parseFormattedTime("11:30", 'decimal-string')).toEqual("11.50");
+        expect(parseFormattedTime("09:05", 'decimal-string')).toEqual("9.08");
+        //expect(parseFormattedTime("67:21", 'decimal-number')).toThrowError("The given time is not valid")
+        expect(parseFormattedTime("02:03", 'decimal-string')).toEqual("2.05")
+        expect(parseFormattedTime("08:03", 'decimal-string')).toEqual("8.05")
+
+        expect(parseFormattedTime("11:57:00", 'decimal-string')).toEqual("11.95");
+        expect(parseFormattedTime("12:07:12", 'decimal-string')).toEqual("12.12");
+        expect(parseFormattedTime("10:51:00", 'decimal-string')).toEqual("10.85");
+        expect(parseFormattedTime("14:12:00", 'decimal-string')).toEqual("14.20");
+        expect(parseFormattedTime("10:37:48", 'decimal-string')).toEqual("10.63");
+        expect(parseFormattedTime("08:25:48", 'decimal-string')).toEqual("8.43");
+        expect(parseFormattedTime("00:00:00", 'decimal-string')).toEqual("0.00");
+        expect(parseFormattedTime("11:52:48", 'decimal-string')).toEqual("11.88");
+        expect(parseFormattedTime("12:27:00", 'decimal-string')).toEqual("12.45");
+        expect(parseFormattedTime("09:54:00", 'decimal-string')).toEqual("9.90");
+        expect(parseFormattedTime("12:37:48", 'decimal-string')).toEqual("12.63");
+        expect(parseFormattedTime("11:30:00", 'decimal-string')).toEqual("11.50");
+        expect(parseFormattedTime("09:04:48", 'decimal-string')).toEqual("9.08");
+        //expect(parseFormattedTime("67:21:00", 'decimal-number')).toThrow(new Error("The given time is not valid"))
+        expect(parseFormattedTime("02:03:00", 'decimal-string')).toEqual("2.05")
+        expect(parseFormattedTime("08:03:00", 'decimal-string')).toEqual("8.05")
+
+        //#endregion
 
     })