+++ time_parsers/time_parsers.js
@@ -2,6 +2,7 @@
 import { toFixed, toFixedFloat } from '../number_parsers/index';
 // Installed Package
 import VDate from '@vieolo/date';
+import { time24Hours, time24HoursWithSeconds } from '@vieolo/regex-library';
 export function parseHourAndMinute(hour, minute, returnType) {
     if (returnType && returnType == 'date') {
         let now = new Date();
@@ -42,4 +43,32 @@ export function parseDecimalTime(decimal, includeSecond) {
     let outputMinute = `${minutes < 10 ? "0" + minutes : minutes}`;
     let outputSecond = `${finalSeconds < 10 ? "0" + finalSeconds : finalSeconds}`;
     return `${outputHour}:${outputMinute}${includeSecond ? ":" + outputSecond : ""}`;
+}
+export function parseMinuteCount(count, returnType) {
+    if (returnType == 'decimal-number')
+        return toFixedFloat(count / 60, 2);
+    else if (returnType == 'decimal-string')
+        return toFixed(count / 60, 2);
+    else if (returnType == 'formatted-hour-minute')
+        return VDate.getTimeFromMinuteCount(count);
+    let hourCount = Math.floor(count / 60);
+    let minuteCount = count % 60;
+    if (returnType == 'formatted-time')
+        return parseHourAndMinute(hourCount, minuteCount, 'string');
+    else
+        return parseHourAndMinute(hourCount, minuteCount, 'vdate');
+}
+export function parseFormattedTime(time, returnType) {
+    if (!time24Hours.test(time) && !time24HoursWithSeconds.test(time))
+        throw new Error("The given time is not valid");
+    let splitted = time.split(':');
+    let minuteCount = (parseInt((splitted[0])) * 60) +
+        parseInt(splitted[1]) +
+        ((splitted.length > 2 && parseInt(splitted[2]) > 29) ? 1 : 0); // Adding one minute if the duration has any seconds
+    let parsed;
+    if (returnType == 'decimal-number')
+        parsed = parseMinuteCount(minuteCount, 'decimal-number');
+    else
+        parsed = parseMinuteCount(minuteCount, 'decimal-string');
+    return parsed;
 }