+++ changelog.md
@@ -1,6 +1,12 @@
 # Change Log
 
 
+### 0.2.0 (2020-06-19)
+
+- Added `parseMinuteCount` function to the time parsers
+- Added `parseFormattedTime` function to the time parsers
+
+
 ### 0.1.0 (2020-06-14)
 
 - Initial Release